mcp-servers/photo-source/pyproject.toml
[project]
name = "photo-source"
version = "0.1.0"
description = "Photo source access MCP server for OpenClaw (Apple Photos, GCS, local)"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "pillow>=10.0",
    "pillow-heif>=0.16",
    "pydantic>=2.0",
]

[project.optional-dependencies]
apple = [
    "osxphotos>=0.68",
    "pyobjc-core>=12.1; sys_platform == 'darwin' and platform_release >= '22.0'",
    "pyobjc-framework-cocoa>=12.1; sys_platform == 'darwin' and platform_release >= '22.0'",
    "pyobjc-framework-coreaudio>=12.1; sys_platform == 'darwin' and platform_release >= '22.0'",
    "pyobjc-framework-coremedia>=12.1; sys_platform == 'darwin' and platform_release >= '22.0'",
    "pyobjc-framework-quartz>=12.1; sys_platform == 'darwin' and platform_release >= '22.0'",
]
gcs = ["google-cloud-storage>=2.14"]
google = [
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "requests>=2.31",
]
all = ["photo-source[apple,gcs,google]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
